Cleanup for tests and factories.
