--- test/services/db_helpers_spec.ex
@@ -1,4 +1,0 @@
-defmodule Janitor.DBHelpersTest do 
-  use Janitor.ModelCase
-  alias Janitor.DBHelpers
-end
